+++ examples/sap_emptyha_simple_example/main.tf
@@ -28,6 +28,7 @@ module "gcp_sap_emptyha" {
   instance_type            = "${var.instance_type}"
   network_tags             = "${var.network_tags}"
   project_id               = "${var.project_id}"
+  subnetwork_project       = "${var.subnetwork_project}"
   region                   = "${var.region}"
   service_account_email    = "${var.service_account_email}"
   boot_disk_size           = "${var.boot_disk_size}"
@@ -40,7 +41,9 @@ module "gcp_sap_emptyha" {
   secondary_zone           = "${var.secondary_zone}"
   sap_vip                  = "${var.sap_vip}"
   sap_vip_secondary_range  = "${var.sap_vip_secondary_range}"
-  public_ip                = "${var.public_ip}"
+  primary_instance_ip      = "${var.primary_instance_ip}"
+  secondary_instance_ip    = "${var.secondary_instance_ip}"
   sap_vip_internal_address = "${var.sap_vip_internal_address}"
+  public_ip                = "${var.public_ip}"
   ip_cidr_range            = "${var.ip_cidr_range}"
 }