+++ examples/sap_hana_simple_example/main.tf
@@ -17,17 +17,17 @@
 provider "google" {
   version     = "~> 2.6.0"
   credentials = "${file("credentials.json")}"
-  region      = "${var.region}"
 }
 
 module "gcp_sap_hana" {
   source                     = "../../modules/sap_hana"
-  subnetwork                 = "${var.subnetwork}"
   linux_image_family         = "${var.linux_image_family}"
   linux_image_project        = "${var.linux_image_project}"
   instance_name              = "${var.instance_name}"
   instance_type              = "${var.instance_type}"
   project_id                 = "${var.project_id}"
+  subnetwork_project         = "${var.subnetwork_project}"
+  subnetwork                 = "${var.subnetwork}"
   region                     = "${var.region}"
   zone                       = "${var.zone}"
   service_account_email      = "${var.service_account_email}"