+++ modules/netweaver/main.tf
@@ -28,6 +28,11 @@ data "template_file" "netweaver" {
   template = "${file("${path.module}/files/startup.sh")}"
 }
 
+data "google_compute_subnetwork" "subnet" {
+  name    = "${var.subnetwork}"
+  project = "${var.subnetwork_project != "" ? var.subnetwork_project : var.project_id}"
+  region  = "${var.region}"
+}
 
 resource "google_compute_disk" "gcp_nw_pd_0" {
   project = "${var.project_id}"
@@ -79,6 +84,7 @@ resource "google_compute_attached_disk" "gcp_nw_attached_pd_2" {
   disk        = "${element(google_compute_disk.gcp_nw_pd_2.*.self_link, count.index)}"
   instance    = "${google_compute_instance.gcp_nw.self_link}"
 }
+
 resource "google_compute_instance" "gcp_nw" {
   project                   = "${var.project_id}"
   name                      = "${var.instance_name}"
@@ -107,7 +113,7 @@ resource "google_compute_instance" "gcp_nw" {
 
   network_interface {
     subnetwork         = "${var.subnetwork}"
-    subnetwork_project = "${var.project_id}"
+    subnetwork_project = "${var.subnetwork_project != "" ? var.subnetwork_project : var.project_id}"
     dynamic "access_config" {
       for_each = [for i in [""] : i if var.public_ip]
       content {}