+++ modules/sap_emptyha/variables.tf
@@ -42,6 +42,10 @@ variable "project_id" {
   description = "The ID of the project in which the resources will be deployed."
 }
 
+variable "subnetwork_project" {
+  description = "The ID of the shared VPC host project in which the service project's will be deployed."
+}
+
 variable "region" {
   description = "Region to deploy the resources. Should be in the same region as the zone."
 }
@@ -103,6 +107,14 @@ variable "public_ip" {
   description = "Determines whether a public IP address is added to your VM instance."
   default     = false
   type        = bool
+}
+
+variable "primary_instance_ip" {
+  description = "Primary instance ip address"
+}
+
+variable "secondary_instance_ip" {
+  description = "Secondary instance ip address"
 }
 
 variable "sap_vip_internal_address" {