+++ modules/sap_hana_ha/variables.tf
@@ -42,6 +42,9 @@ variable "project_id" {
   description = "The ID of the project in which the resources will be deployed."
 }
 
+variable "subnetwork_project" {
+  description = "The ID of the shared VPC host project in which the service project's will be deployed."
+}
 
 variable "region" {
   description = "Region to deploy the resources. Should be in the same region as the zone."
@@ -183,10 +186,19 @@ variable "public_ip" {
   type        = bool
 }
 
+variable "primary_instance_ip" {
+  description = "Primary instance ip address"
+}
+
+variable "secondary_instance_ip" {
+  description = "Secondary instance ip address"
+}
+
 variable "sap_vip_internal_address" {
   description = "Name of static IP adress to add to the instance's access config."
 }
-
+/*
 variable "ip_cidr_range" {
   description = "ip cidr range for secondary ip ranges"
 }
+*/