+++ modules/sap_hana_ha_svpc/files/startup_secondary.sh
@@ -1,5 +1,0 @@
-#!/bin/bash
-
-curl -s https://storage.googleapis.com/sapdeploy/dm-templates/sap_hana_ha/startup_secondary.sh >> startup_secondary.sh
-chmod +x startup_secondary.sh
-source startup_secondary.sh